position now gotten from the body, and moving works by applying force to the body

+++ rocc/core/src/main/java/edu/chl/rocc/core/Model/RoCCModel.java
@@ -73,11 +73,11 @@ public class RoCCModel {
         player.move(dir);
     }
 
-    public int getCharacterXPos(){
+    public float getCharacterXPos(){
         return player.getCharacterXPos();
     }
 
-    public int getCharacterYPos(){
+    public float getCharacterYPos(){
         return player.getCharacterYPos();
     }
